split Rest Client tests into multiple files
